restructure a few components
