+++ src/App.js
@@ -6,45 +6,32 @@ import "./css/Graph.css";
 
 import "./css/Media.css";
 
-//built-in components
-import ReactLoading from "react-loading";
-import Twemoji from "react-emoji-render";
-
-//components
-import Statusblock from "./components/Statusblock";
-import { CountryHeader } from "./components/CountryHeader";
-import TimelineHeader from "./components/TimelineHeader";
-import StartHeader from "./components/StartHeader";
-import StatLoader from "./components/StatLoader";
+import {
+  StartHeader,
+  SelectCountrySection,
+  CovidStatsByCountrySection,
+  CovidStatsByCountryTableSection,
+} from "./components";
 
 //custom data and function modules
-import { worldImg, countryImg, pieSizeOpts, pieLegendOpts } from "./data/data";
-import { structureCountryData, structurePieData } from "./data/functions";
-import { formatHistoricalData, formatWeeklyData } from "./data/functions";
-import { barStyleGetter } from "./data/functions";
-
-//custom graph custom styles
-import { globalLegendOpts, lineOpts } from "./data/data";
-import ReactSelectCustomComponents from "./helpers/ReactSelectCustomComponents";
+import { worldImg, countryImg } from "./data/data";
+import {
+  structureCountryData,
+  structurePieData,
+  formatHistoricalData,
+  formatWeeklyData,
+} from "./data/functions";
 
 //custom api_functions modules
 import {
+  getCountryNames,
   getCountryCovidData,
   getWorldWideCovidData,
-} from "./data/api_functions";
-import {
   getLastWeekDataOfWorld,
   getLastWeekDataByCountry,
 } from "./data/api_functions";
 
-import { getCountryNames } from "./data/api_functions";
-
 import defaultCountryOption from "./constants/defaultObjects";
-
-import AsyncSelect from "react-select/async";
-
-//package components from react-chartjs-2 npm package
-import { Pie, Bar, Line, HorizontalBar } from "react-chartjs-2";
 
 function App() {
   const [countryList, setCountryList] = useState([]);
@@ -73,48 +60,6 @@ function App() {
   const viewTypeChange = (viewTypeOption) => {
     setViewsSelect(viewTypeOption?.value);
   };
-
-  function renderViewType(countryWeekCases) {
-    if (histLoading)
-      return (
-        <ReactLoading
-          type="bars"
-          color={"black"}
-          height={"70px"}
-          width={"70px"}
-        />
-      );
-
-    if (noHistory)
-      return (
-        <div className="history-error">
-          No timeline data available <Twemoji text=":/" />
-        </div>
-      );
-
-    const visualType =
-      viewsSelect === "Horizontal Bar" ? (
-        <HorizontalBar
-          data={countryWeekCases}
-          options={barStyleGetter(true)}
-          legend={globalLegendOpts}
-        />
-      ) : viewsSelect === "Line" ? (
-        <Line
-          data={countryWeekCases}
-          options={lineOpts}
-          legend={globalLegendOpts}
-        />
-      ) : (
-        <Bar
-          data={countryWeekCases}
-          options={barStyleGetter(false)}
-          legend={globalLegendOpts}
-        />
-      );
-
-    return visualType;
-  }
 
   const loadCountryNamesList = () =>
     getCountryNames().then((response) => {
@@ -220,90 +165,34 @@ function App() {
   return (
     <>
       <StartHeader />
-      <div id="country-selection">
-        <div id="country-select-heading">View Global / Local Information</div>
-        <div id="countries_select">
-          <AsyncSelect
-            cacheOptions
-            defaultOptions
-            isSearchable={false}
-            loadOptions={loadCountryNamesList}
-            value={countrySelected}
-            onChange={onSelectChange}
-            components={ReactSelectCustomComponents}
-          />
-        </div>
-      </div>
-
-      <CountryHeader
-        loading={loading}
-        countryFlag={countryFlag}
-        countryName={countryTitle}
-        statError={statError}
-        imgStyle={countryName !== "World" ? countryImg : worldImg}
+      <SelectCountrySection
+        loadCountryNamesList={loadCountryNamesList}
+        countrySelected={countrySelected}
+        onSelectChange={onSelectChange}
       />
-
-      <div className="country-info-container">
-        {statError ? (
-          <div className="stat-error">
-            <div>
-              Something went wrong <Twemoji text=":'(" />
-            </div>
-          </div>
-        ) : null}
-
-        {countryData && !loading ? (
-          countryData.map((curr, idx) => (
-            <Statusblock key={idx + 1} packet={curr} />
-          ))
-        ) : loading ? (
-          <StatLoader />
-        ) : null}
-
-        {!statError ? (
-          <div className="chart-container">
-            {pieData && !loading ? (
-              <Pie
-                data={pieData}
-                width={420}
-                height={420}
-                legend={pieLegendOpts}
-                options={pieSizeOpts}
-              />
-            ) : (
-              <div className="chart-loader">
-                <ReactLoading
-                  type={"bars"}
-                  color={"black"}
-                  height={"70px"}
-                  width={"70px"}
-                />
-              </div>
-            )}
-          </div>
-        ) : null}
-      </div>
-
-      <TimelineHeader
+      <CovidStatsByCountrySection
+        loading={loading}
+        countryData={countryData}
+        countryFlag={countryFlag}
+        countryName={countryName}
+        countryTitle={countryTitle}
+        statError={statError}
+        countryImg={countryImg}
+        worldImg={worldImg}
+        pieData={pieData}
+      />
+      <CovidStatsByCountryTableSection
         graphError={graphError}
         noHistory={noHistory}
-        viewChange={viewTypeChange}
-        daysChange={daysAmntChange}
+        viewTypeChange={viewTypeChange}
+        daysAmntChange={daysAmntChange}
         daysSelect={daysSelect}
         viewsSelect={viewsSelect}
+        countryWeekCases={countryWeekCases}
+        countryData={countryData}
+        countryList={countryList}
+        histLoading={histLoading}
       />
-
-      <div className="graph-section">
-        <div className="graph-container">
-          {countryWeekCases && countryData && countryList.length !== 0 ? (
-            renderViewType(countryWeekCases)
-          ) : graphError ? (
-            <div className="graph-error">
-              Something went wrong <Twemoji text=":'(" />
-            </div>
-          ) : null}
-        </div>
-      </div>
     </>
   );
 }
