+++ src/components/index.jsx
@@ -1,0 +1,11 @@
+import StartHeader from "./StartHeader";
+import SelectCountrySection from "./SelectCountrySection";
+import CovidStatsByCountrySection from "./CovidStatsByCountrySection";
+import CovidStatsByCountryTableSection from "./CovidStatsByCountryTableSection";
+
+export {
+  StartHeader,
+  SelectCountrySection,
+  CovidStatsByCountrySection,
+  CovidStatsByCountryTableSection,
+};
